Ispravka greske za tip podatka Int -> Float

+++ ForecastMVVM/app/src/main/java/com/example/sale1996/forecastmvvm/data/db/entity/Current.kt
@@ -15,7 +15,7 @@ data class Current(
     val humidity: Int,
     @SerializedName("observation_time")
     val observationTime: String,
-    val precip: Int,
+    val precip: Float,
     val pressure: Int,
     val temperature: Int,
     @SerializedName("uv_index")
