+++ ForecastMVVM/app/src/main/java/com/example/sale1996/forecastmvvm/data/db/entity/CurrentWeather.kt
@@ -14,7 +14,7 @@ data class CurrentWeather(
     val humidity: Int,
     @SerializedName("observation_time")
     val observationTime: String,
-    val precip: Int,
+    val precip: Float,
     val pressure: Int,
     val temperature: Int,
     @SerializedName("uv_index")
