+++ ForecastMVVM/app/src/main/java/com/example/sale1996/forecastmvvm/data/db/unitlocalized/UnitSpecificCurrentWeatherEntry.kt
@@ -7,7 +7,7 @@ interface UnitSpecificCurrentWeatherEntry {
     val feelslike: Int
     val humidity: Int
     val observationTime: String
-    val precip: Int
+    val precip: Float
     val pressure: Int
     val temperature: Int
     val uvIndex: Int
